Change amount repr to double

+++ src/main/java/seedu/address/model/epiggy/Savings.java
@@ -5,7 +5,7 @@ package seedu.address.model.epiggy;
  * Guarantees: details are present and not null, field values are validated, immutable.
  */
 public class Savings {
-    private float savings;
+    private double savings;
 
     public Savings() {
         this.savings = 0.00f;
@@ -15,15 +15,15 @@ public class Savings {
         this.savings = savings.getSavings();
     }
 
-    public float getSavings() {
+    public double getSavings() {
         return savings;
     }
 
-    public void addSavings(float increment) {
+    public void addSavings(double increment) {
         savings += increment;
     }
 
-    public void deductSavings(float decrement) {
+    public void deductSavings(double decrement) {
         savings -= decrement;
     }
